v4.3.1: Presumably fixes backward compatibility issue with processCss
This issue:
https://github.com/css-modules/css-modules-require-hook/issues/115

The version v4.2.2 introduced a breaking change for people relying on
`processCss` option. The actual reason was that since that version
"postcss-modules-resolve-imports" combined separate CSS into single CSS
prior to triggering processCss callback, if any was provided. Up to
4.2.1 the behavior was different, and processCss was called on each
individual file.

This patch, if `processCss` callback is provided, with a help of
additional simple plugin, records separate compiled CSS files, and then
passes them to `processCss` callback individually, with the same order
as before.

As of now, I am not sure it solves the issue in all corner cases, but at
least with this patch `babel-plugin-css-modules-transform` passes all
its functional tests when upgraded to the v4.3.1 of
`css-modules-require-hook`.