v4.5.0: Updates classname generation logic
After this change:
- Without `generateScopedName` option the hook generates the same class
names as before;
- With `generateScopedName` option set to string, the hook generates the
same class names as `css-loader@5.0.1`. The direct dependency on
`css-loader` had to be dropped, asthe latest `css-loader` does not allow
the hack used previously to ensure the same class name generation. Now,
the necessary code is clonned into the hook codebase.
- With `generateScopedName` option set to function, the hook uses that
function to generate class names.

The change is marked as a minor version update, because I am not 100%
sure it does not alter the names generated in second case.